requirements.txt
joblib==1.5.2
numpy==2.3.3
torch==2.8.0
torch
scikit-learn
pandas
fastapi
uvicorn
streamlit
requests
biopython